FP-18: Moved notes and history to new components, CaseNotes and CaseActivity (named per guidelines on jira). Because timestamps have yet to be implemented, created quick and dirty dummy data for now. Added comments about next steps (after FP-18)
